backend/heka/requirements.txt
asgiref==3.4.1
asttokens==2.0.8
certifi==2021.5.30
colorama==0.4.5
Django==3.2
django-cors-headers==3.10.1
djangorestframework==3.14.0
executing==1.1.1
httplib2==0.20.4
i==2.1.0
install==1.3.5
pip==21.2.2
psycopg2-binary==2.9.4
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.1
pytz==2021.3
PyYAML==6.0
setuptools==58.0.4
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.1.1
wheel==0.37.1
uritemplate
drf-yasg
django-autoslug